WIP #62 Fixed up observable issues on Podcast Archive page
