+++ src/DNI.Web/src/app/components/shared/pager/pager.component.html
@@ -1,5 +1,4 @@
-<div class="dni-pager"
-     *ngIf="totalItems > 0">
+<div class="dni-pager">
     <pagination [totalItems]="totalItems"
                 [itemsPerPage]="itemsPerPage"
                 [(ngModel)]="currentPage"
