+++ src/DNI.Web/src/app/services/show/get-shows-response.ts
@@ -15,4 +15,22 @@ export class GetShowsResponse {
         return Object.keys(this.showKeywords)
             .map(s => new ShowKeyword(s, this.showKeywords[s]));
     }
+
+    public getTopKeywords(top: number = 10): ShowKeyword[] {
+        return this.getKeywords()
+            .sort((a, b) => {
+                const countSort = (a.count * -1) - (b.count * -1);
+                let keywordSort: number;
+                if (a.keyword < b.keyword) {
+                    keywordSort = -1;
+                } else if (a.keyword > b.keyword) {
+                    keywordSort = 1;
+                } else {
+                    keywordSort = 0;
+                }
+
+                return countSort || keywordSort;
+            })
+            .slice(0, top);
+    }
 }
